Create moimoi.command package
Command classes are located on the same directory level as other classes
used by MoiMoi.

This unorganized and cluttered project structure makes the codebase
harder to maintain as the number of classes increases.

Let's create a package named "moimoi.command" and move all
command-related classes into this package.

This improves code maintainability by grouping related classes together.
The codebase also becomes cleaner and easier to navigate.
