+++ src/main/java/moimoi/command/CommandEnum.java
@@ -1,3 +1,5 @@
+package moimoi.command;
+
 public enum CommandEnum {
     TODO,
     DEADLINE,
