+++ src/main/java/moimoi/command/UnmarkCommand.java
@@ -1,8 +1,16 @@
-public class DeleteCommand extends Command {
+package moimoi.command;
+
+import moimoi.Storage;
+import moimoi.TaskList;
+import moimoi.Ui;
+import moimoi.exception.InvalidIndexException;
+import moimoi.task.Task;
+
+public class UnmarkCommand extends Command {
 
     String indexString;
 
-    public DeleteCommand(String indexString) {
+    public UnmarkCommand(String indexString) {
         super(false);
         this.indexString = indexString;
     }
@@ -12,9 +20,8 @@ public class DeleteCommand extends Command {
         try {
             int index = Integer.parseInt(this.indexString);
             Task task = tasks.get(index);
-            tasks.delete(index);
-            ui.showCompletionMessage("Aju nice! I've got rid of this task: " + task.stringUI()
-                    + "\nWe have " + tasks.size() + " tasks left in the bag~");
+            task.unmark();
+            ui.showCompletionMessage("Oof, it's OK! Let's get it done soon ;)\n" + task.stringUI());
         } catch (NumberFormatException | IndexOutOfBoundsException e) {
             throw new InvalidIndexException();
         }
